♻️ refactor: Viewers를 공통 컴포넌트로 변경 - viewer 뿐 아니라 vote에서도 쓰이는 컴포넌트
